Remove Copyable requirement from RandomNumberGenerator protocol
We should allow RNGs to be ~Copyable. For one thing, they simply don't need to be copyable. For another, it's entirely plausible that it would be desirable to bind an RNG to a unique resource. It's entirely arguable that the SystemRNG should be ~Copyable, but I'll consider that under a separate PR.