[utils][presets] Build libc++ for LLDB on linux
We pass `-stdlib=libc++` in the test-suite to
Clang for many of our tests because we want to
explicilty test against libc++ type layouts.
However, if we don't do this against a newly built
libc++ we risk testing against old/unexpected layouts
or if libc++ isn't available on the system, falling
back to the system libstdc++.

This patch adds libcxx as an explicit target in the preset
so the tests build against a fresh libc++ (as we do for
our other LLDB presets too).