[concurrency] Add support for swift_task_isOnMainActor, _isOnMainActor, _runOnMainActor.
swift_task_isOnMainActor is a C++ API that checks if one is on the main actor:

1. In synchronous code that doesn't use swift concurrency, returning
   Thread::isOnMainThread().

2. In synchronous or asynchronous code running within Swift concurrency,
   ExecutorTrackingInfo is used to determine the current executor and compare that
   against the main executor.

Notice that swift_task_isOnMainActor is new ABI, so we want to on Darwin (the
only platform with a stable ABI) be able to back deploy this as far as
possible. Luckily on other platforms this is not an actual issue since they do
not have a stable ABI.

On Darwin, we back deploy using the API _isOnMainActor. _isOnMainActor on
non-Darwin platforms and Darwin platforms that are new enough to have
swift_task_isOnMainActor, just call that. Otherwise, we are on an older Darwin
system and we need to use posix and dispatch APIs to implement the same general
algorithm by first determining if we are called from code within Swift
Concurrency by checking if we have an async task running and then:

1. If there isn't an async task running, then we know that we are in a
   synchronous context outside of Swift concurrency. We then just return
   pthread_main_np to return true if we are on the main thread.

2. If there is an async task, we need to use dispatch APIs to determine if we
   are running on the main queue or a child queue of the main queue. To do that,
   we first grab the main queue by grabbing it from Swift's main executor via
   some bitcasting. Then we use dispatch_get_current_queue() to get the current
   queue working in dispatch and compare it against the main queue. If they
   equal, then we know that we are on the main queue. Sadly, we cannot just do
   that since we could have child queues of the main queue, which are considered
   to be on the main actor. So in that case, we unconditionally use
   dispatch_queue_set_specific to set onto the main actor the address of the
   main queue repeated as a key value pair. Then we check the potential child
   queue for the address of the queue and see if we get back the queue using
   dispatch_queue_set_specific. This is taking advantage of the property that
   dispatch_queue_{get,set}_specific automatic ally handle child tasks for the
   caller.

This is the reason why I am implementing the above two entry points. This
function is a nonisolated synchronous function that takes in a main actor
isolated synchronous closure. When run, the code checks dynamically if we are on
the main actor and in such a case just runs the code synchronously. Otherwise,
if we are not on the main actor, a new Task is created and the MainActor
isolated function is run on it. NOTE: We only allow this for closures that do
not take arguments or results (a pure callback).