[Concurrency] Fix racy test, too sensitive on println ordering
The test only checks Task done and then "deinit" happening, however
there is another print that may happen in between. Since it's racy which
one comes first, on slower instances -- like simulators, it could get in
the way.

Remove the println we're not even checking for but that breaks the
CHECK-NEXT

Resolves rdar://137221764