[Windows] Build swift-format using SwiftPM from build.ps1 and run tests
The primary motivation is to run swift-format tests on Windows using 'swift test'.

Since we now have infrastructure for testing swift-format using SwiftPM, we should also use it to build swift-format to be included in the toolchain because (1) it means we don't need to build SwiftFormat twice, (2) we test the same build that is also included in the toolchain and (3) it allows us to remove the CMake build of swift-format.