[DebugInfo] Emit types with @_originallyDefinedIn under original module
LLDB does not have enough information to find types with the
@_originallyDefinedIn attribute in reflection metadata and in
swiftmodules.

When emitting debug info for such a type, emit it under the module where
the type was originally defined, but in the mangled name emit the
current ABI module name.

This way LLDB has enough information to locate the type in both places.

rdar://137146961