Add a flag to dump the AST as JSON.
The new `-dump-ast-format` flag can be supplied when invoking the compiler
in `-dump-ast` mode. Valid values for this flag are `default` (the default
pseudo-S-expression format) and `json`.

This is meant to be used by clients who want to do large-scale semantic
analysis of type-checked Swift code. It is written with the understanding
that there are no guarantees of stability for the semantic AST in the
compiler, so clients must read the version information encoded in the
JSON output and adjust their behavior accordingly.