[build-script] Enable new parser validation by default
Widen the parser validation coverage.
This encourages the developers to implement the new grammar in
SwiftParser in addition to the old C++ parser.

In case  there's an unavoidable reason not to implement in SwiftParser,
disable the validation for each test case by passing a frontend option
'-disable-experimental-parser-round-trip'. But please file an issue
and add FIXME like:

  // FIXME: Remove '-disable-experimental-parser-round-trip' (rdar://xxx).