Sema: Fix source ranges for string literals in TypeRefinementContext.
Since string literals are only a single token from the perspective of the
parser, the source range for a pattern binding decl that is initialized with a
string literal expression only extends to the beginning of the string literal.
To ensure the TypeRefinementContext for a pattern binding decl includes the
entire contents of the init expression we must calculate the character source
range of the decl instead.

Resolves rdar://110952225 and https://github.com/swiftlang/swift/issues/77050.