[build] Make the RegexParser requirement to build ASTGen explicit
This makes the requirement to have RegexParser build enabled more obvious to someone not familiar with ASTGen.

This also removes the dummy target that was triggering build failures: certain Swift code in ASTGen was using `#if canImport(_CompilerRegexParser)` to determine whether RegexParser is available, and the dummy target was causing this to always resolve to true. That obscured the underlying reason for compilation errors:
```
ASTGen/Sources/ASTGen/Regex.swift:51:30: error: cannot find 'swiftCompilerLexRegexLiteral' in scope
```